chat: An incredibly basic chat client using UDP/IP and TCP/IP sockets
Clients connect to the monitor using UDP to request a list of
known rooms (the server informs the chat federation server
when there is a new room available and removes it when
its quits). Then they can connect to any one of those
servers and their room and post messages to them.

Kind of like IRC, I guess.